% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bbb}
\alias{bbb}
\title{A toy data}
\format{
A data frame with 86 rows and 2 variables:
\describe{
  \item{T}{finding by T; the method being tested}
  \item{R}{finding by R; the imperfect reference method}
}
}
\source{

}
\usage{
bbb
}
\description{
A toy dataset with results of a test T and an imperfect reference R.
0 indicates absence of a disease being detected, 1 presence of the disease.
}
\keyword{datasets}
