% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{bounds}
\alias{bounds}
\title{Bounds on true sensitivity, specificity, PPV, NPV of a new test}
\usage{
bounds(TR, alphaR, betaR)
}
\arguments{
\item{TR}{concordance table of the test T and the reference R; as in Emerson et al. (2017)}

\item{alphaR}{sensitivity of R (wrt the gold standard)}

\item{betaR}{specificity of R (wrt the gold standard)}
}
\value{
list of bounds:
\describe{
  \item{minaT}{minimal value of sensitivity of the tested method}
  \item{maxaT}{maximal value of sensitivity}
  \item{minbT}{minimal value of specificity}
  \item{maxbT}{maximal value of specificity}
  \item{minPPV}{minimal value of PPV}
  \item{maxPPV}{maximal value of PPV}
  \item{minNPV}{minimal value of NPV}
  \item{maxNPV}{maximal value of NPV}
}
}
\description{
\code{bounds} computes Emerson et al. (2017) bounds on true sensitivity and specificity of a test method wrt an imperfect reference method, as well as (trivially) implied bounds on PPV, NPV.
}
\details{
The TR table of this form is expected:
\tabular{ccc}{
T vs R \tab 1 \tab 0 \cr
1 \tab n11 \tab n10 \cr
0 \tab n01 \tab n00 \cr
}

If the disease prevalence theta, as computed by the formula on p. 6 in Emerson et al. (2017) is not in (0,1), list of NAs is returned.
}
\examples{
## TR concordance table must have the same form as in Emerson et al. (2017).
##       T\R 1 0
##       1
##       0
TR = c(7, 1, 2, 29)
dim(TR) = c(2, 2)
TR
## Hence, it happened in 2 cases
## that the test method reported '1' (presence of disease)
## and the reference method reported '0' (absence of disease)

alphaR = 0.98
betaR = 0.96

bounds(TR, alphaR, betaR)

}
\references{
Emerson S.C., Waikar S.S., Fuentes C., Bonventre J.V., Betensky R.A. (2017). Biomarker validation with an imperfect reference: Issues and bounds, Statistical Methods in Medical Research, 27(10):2933-2945.
}
